requirements.txt
streamlit==1.28.0
openai==1.3.0
instructor==0.4.8
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0
plotly==5.17.0
scikit-learn==1.3.0